Add files via upload
This R-code describes the complete path of a fertilizer particle of a blend (or one fertilizer type), from spreader hopper to spinning disk to air to ground surface. The model complements the analytical approach taken by Villette et al. (2005) to the motion of a fertilizer particle on a rotating disk and the Hybrid Centrifugal Spreading Model (HCSM) published by Villette et al. (2017)